ref: refactor stuff
