+++ src/main/java/org/example/handlers/rest/LoginHandler.java
@@ -9,12 +9,12 @@ import com.amazonaws.services.lambda.runtime.events.APIGatewayV2HTTPResponse;
 import com.google.gson.Gson;
 import org.example.constants.StatusCodes;
 import org.example.entities.User;
+import org.example.entities.session.SessionService;
 import org.example.exceptions.Unauthorized;
 import org.example.models.requests.LoginRequest;
 import org.example.models.requests.SessionRequest;
 import org.example.models.responses.rest.LoginResponseBody;
 import org.example.services.LoginService;
-import org.example.services.SessionService;
 import org.example.utils.ValidateObject;
 
 public class LoginHandler
