+++ src/main/java/org/example/handlers/websocket/JoinGameHandler.java
@@ -11,8 +11,8 @@ import java.util.Optional;
 import org.example.constants.StatusCodes;
 import org.example.entities.Game;
 import org.example.entities.Player;
-import org.example.entities.Stats;
 import org.example.entities.User;
+import org.example.entities.stats.Stats;
 import org.example.enums.Action;
 import org.example.enums.GameMode;
 import org.example.models.requests.JoinGameRequest;
