+++ src/main/java/org/example/services/JoinGameService.java
@@ -8,8 +8,8 @@ import lombok.AllArgsConstructor;
 import org.bson.conversions.Bson;
 import org.example.constants.ChessConstants;
 import org.example.entities.Game;
-import org.example.entities.Stats;
 import org.example.entities.User;
+import org.example.entities.stats.Stats;
 import org.example.enums.GameStatus;
 import org.example.enums.TimeControl;
 import org.example.utils.MongoDBUtility;
