+++ src/main/java/org/example/services/LogoutService.java
@@ -3,6 +3,7 @@ package org.example.services;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.NoArgsConstructor;
+import org.example.entities.session.SessionService;
 import org.example.enums.ResultReason;
 import org.example.exceptions.InternalServerError;
 import org.example.exceptions.NotFound;
