+++ src/test/java/org/example/handlers/getGameState/GetGameStateTest.java
@@ -16,8 +16,8 @@ import java.util.Optional;
 import org.example.constants.StatusCodes;
 import org.example.entities.Game;
 import org.example.entities.Player;
-import org.example.entities.Stats;
 import org.example.entities.User;
+import org.example.entities.stats.Stats;
 import org.example.enums.GameStatus;
 import org.example.enums.TimeControl;
 import org.example.handlers.rest.GetGameStateHandler;
