+++ src/test/java/org/example/handlers/logout/LogoutHandlerTest.java
@@ -9,12 +9,14 @@ import java.util.Map;
 import java.util.Optional;
 import org.example.constants.StatusCodes;
 import org.example.entities.*;
+import org.example.entities.session.Session;
+import org.example.entities.session.SessionService;
+import org.example.entities.stats.Stats;
 import org.example.enums.GameStatus;
 import org.example.enums.TimeControl;
 import org.example.handlers.rest.LogoutHandler;
 import org.example.services.GameStateService;
 import org.example.services.LogoutService;
-import org.example.services.SessionService;
 import org.example.utils.MockContext;
 import org.example.utils.MongoDBUtility;
 import org.example.utils.socketMessenger.SocketSystemLogger;
