+++ src/test/java/org/example/handlers/makeMove/MakeMoveHandlerTest.java
@@ -14,8 +14,8 @@ import org.example.constants.ChessConstants;
 import org.example.constants.StatusCodes;
 import org.example.entities.Game;
 import org.example.entities.Player;
-import org.example.entities.Stats;
 import org.example.entities.User;
+import org.example.entities.stats.Stats;
 import org.example.enums.Action;
 import org.example.enums.GameStatus;
 import org.example.enums.TimeControl;
