+++ src/test/java/org/example/handlers/register/RegisterHandlerTest.java
@@ -11,8 +11,8 @@ import java.util.Map;
 import java.util.Optional;
 import org.bson.conversions.Bson;
 import org.example.constants.StatusCodes;
-import org.example.entities.Stats;
 import org.example.entities.User;
+import org.example.entities.stats.Stats;
 import org.example.handlers.rest.RegisterHandler;
 import org.example.models.requests.RegisterRequest;
 import org.example.services.RegisterService;
