+++ src/test/java/org/example/handlers/stats/StatsHandlerTest.java
@@ -7,8 +7,8 @@ import com.amazonaws.services.lambda.runtime.events.APIGatewayV2HTTPEvent;
 import com.amazonaws.services.lambda.runtime.events.APIGatewayV2HTTPResponse;
 import java.util.Map;
 import org.example.constants.StatusCodes;
-import org.example.entities.Stats;
 import org.example.entities.User;
+import org.example.entities.stats.Stats;
 import org.example.handlers.rest.StatsHandler;
 import org.example.utils.MockContext;
 import org.example.utils.MongoDBUtility;
